ensure more consistency across all dashboard components
